refactor: separate root and match command in two files
